fix: resolve all remaining CI compile errors
- Add PasswordVisualTransformation/VisualTransformation imports to ConfigScreen.kt
- Remove AppNavGraph.kt (conflict with NavGraph.kt)
- Fix HomeScreen.kt: replace .weight(1f) with .fillMaxWidth(), stub HomeStatusBar
- Fix AgentService.kt: use android.R.drawable.ic_dialog_info
- Fix SettingsRepository.kt: add DataStore<Preferences> generic + import
- Add datastore-preferences/datastore-core to build.gradle.kts
- Fix PythonRuntime.kt core: add StreamBridge import
- Fix ExpandedOverlayView.kt: dp(0.5f).toInt() for setStroke
- Delete ConfigScreenV2.kt (conflict with ConfigScreen.kt)
- Fix ChatScreen.kt: add @OptIn(ExperimentalMaterial3Api)
- Fix VoiceInputButton.kt: add missing background import
- Fix AgentDaemonV2.kt: replace wildcard imports
- Fix AgentDaemon.kt: remove android.os.Process import
- Fix AppBootstrapper.kt: add default repoUrl parameter
- Fix MiniOverlayView.kt: add GradientDrawable import
- Fix TermuxBootstrap.kt: add termuxRoot property

+++ app/src/main/java/com/pocketagent/app/overlay/ExpandedOverlayView.kt
@@ -52,9 +52,9 @@ class ExpandedOverlayView(
         // 整体背景
         background = GradientDrawable().apply {
             shape = GradientDrawable.RECTANGLE
-            cornerRadius = dp(12f)
+            cornerRadius = dp(12).toFloat()
             setColor(Color.parseColor("#E61A1A2E"))
-            setStroke(dp(0.5f), Color.parseColor("#44FFFFFF"))
+            setStroke(dp(0.5f).toInt(), Color.parseColor("#44FFFFFF"))
         }
 
         layoutParams = ViewGroup.LayoutParams(
@@ -70,7 +70,7 @@ class ExpandedOverlayView(
             background = GradientDrawable().apply {
                 shape = GradientDrawable.RECTANGLE
                 cornerRadii = floatArrayOf(
-                    dp(12f), dp(12f), dp(12f), dp(12f), // top
+                    dp(12).toFloat(), dp(12).toFloat(), dp(12).toFloat(), dp(12).toFloat(), // top
                     0f, 0f, 0f, 0f                      // bottom
                 )
                 setColor(Color.parseColor("#331A1A2E"))
